Refactored the NotifyCounts component out of the navbar.jsx file into its own file
