+++ web/react/components/navbar.jsx
@@ -1,7 +1,6 @@
 // Copyright (c) 2015 Spinpunch, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-var utils = require('../utils/utils.jsx');
 var client = require('../utils/client.jsx');
 var AsyncClient = require('../utils/async_client.jsx');
 var UserStore = require('../stores/user_store.jsx');
@@ -10,53 +9,11 @@ var TeamStore = require('../stores/team_store.jsx');
 
 var UserProfile = require('./user_profile.jsx');
 var MessageWrapper = require('./message_wrapper.jsx');
+var NotifyCounts = require('./notify_counts.jsx');
 
 var Constants = require('../utils/constants.jsx');
 var ActionTypes = Constants.ActionTypes;
 var AppDispatcher = require('../dispatcher/app_dispatcher.jsx');
-
-function getCountsStateFromStores() {
-    var count = 0;
-    var channels = ChannelStore.getAll();
-    var members = ChannelStore.getAllMembers();
-
-    channels.forEach(function setChannelInfo(channel) {
-        var channelMember = members[channel.id];
-        if (channel.type === 'D') {
-            count += channel.total_msg_count - channelMember.msg_count;
-        } else if (channelMember.mention_count > 0) {
-            count += channelMember.mention_count;
-        } else if (channelMember.notify_level !== 'quiet' && channel.total_msg_count - channelMember.msg_count > 0) {
-            count += 1;
-        }
-    });
-
-    return {count: count};
-}
-
-var NotifyCounts = React.createClass({
-    componentDidMount: function() {
-        ChannelStore.addChangeListener(this.onListenerChange);
-    },
-    componentWillUnmount: function() {
-        ChannelStore.removeChangeListener(this.onListenerChange);
-    },
-    onListenerChange: function() {
-        var newState = getCountsStateFromStores();
-        if (!utils.areStatesEqual(newState, this.state)) {
-            this.setState(newState);
-        }
-    },
-    getInitialState: function() {
-        return getCountsStateFromStores();
-    },
-    render: function() {
-        if (this.state.count) {
-            return <span className='badge badge-notify'>{this.state.count}</span>;
-        }
-        return null;
-    }
-});
 
 function getStateFromStores() {
     return {
